requirements.txt
absl-py==0.11.0
apache-airflow==2.1.2
seaborn==0.11.1
